move createSegmentFromBoundary into its own file
